fix(web): break sales-demo import cycle
